refactor: moved api.spec and testHelper into testFolder
