+++ apps/miranum-cli/src/app/deployment/deployment.ts
@@ -39,6 +39,8 @@ export class Deployment {
                 await this.deploy(file.file, file.type, target);
             } catch (err) {
                 console.log(colors.red.bold("FAILED ") + `deploying ${file.file.name} with -> ${err}`);
+                // todo the deployment should continue and fail in the end, not here (Übergangslösung)
+                return Promise.reject(colors.red.bold("FAILED ") + `deploying Artifacts in ${path}`);
             }
         }
     }
