+++ apps/miranum-cli/src/app/deployment/deployments.spec.ts
@@ -1,7 +1,7 @@
 import { Deployment } from "./deployment";
 import {Artifact, createMiranumCore, MiranumDeploymentPlugin} from "@miranum-ide/miranum-core";
 import * as colors from "colors";
-import {filesToDeploy, pathToProject} from "../shared/testHelpers";
+import {filesToDeploy, pathToProject} from "../../../tests/testHelpers";
 
 const sampleTarget = "local";
 
@@ -28,8 +28,8 @@ describe("deployArtifact", () => {
     for(const file of filesToDeploy) {
         it(`${file.type} should work`, async () => {
             const logSpy = jest.spyOn(console, 'log');
-            await deployment.deployArtifact(file.path, file.type, sampleTarget);
-
+            await expect(deployment.deployArtifact(file.path, file.type, sampleTarget))
+                .resolves.not.toThrow();
             expect(logSpy).toHaveBeenCalledWith(colors.green.bold("DEPLOYED ") + file.nameExt + " to environment " + sampleTarget);
         });
     }
@@ -43,12 +43,12 @@ describe("deployArtifact", () => {
 describe("deployAllArtifacts", () => {
     it("should work", async () => {
         const logSpy = jest.spyOn(console, 'log');
-        await deployment.deployAllArtifacts(pathToProject, sampleTarget);
+        await expect(deployment.deployAllArtifacts(pathToProject, sampleTarget))
+            .resolves.not.toThrow();
 
         for (const file of filesToDeploy) {
             expect(logSpy).toHaveBeenCalledWith(colors.green.bold("DEPLOYED ") + file.nameExt + " to environment " + sampleTarget);
         }
-
     });
 
     it("should raise an error", async () => {
