+++ package.json
@@ -62,6 +62,7 @@
     "colors": "^1.4.0",
     "commander": "^9.4.0",
     "core-js": "^3.6.5",
+    "madge": "^5.0.1",
     "primeicons": "^6.0.1",
     "primereact": "^8.7.2",
     "react": "18.2.0",
